Linux Driver: move use space header
To avoid conflict with Azure linux header
